% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_density.R
\name{calc_density}
\alias{calc_density}
\title{Calculate Density Estimation}
\usage{
calc_density(
  data = NULL,
  trim = TRUE,
  bw = "nrd0",
  adjust = 1,
  kernel = "gaussian",
  n = 512
)
}
\arguments{
\item{data}{A data frame containing the dataset for density estimation.}

\item{trim}{Logical value indicating whether to trim values outside the data range.}

\item{bw}{Bandwidth for density estimation (default is "nrd0").}

\item{adjust}{A bandwidth adjustment factor (default is 1).}

\item{kernel}{The kernel function to use (default is "gaussian").}

\item{n}{Number of grid points for density estimation (default is 512).}
}
\value{
A data frame containing density estimation results.
}
\description{
This function calculates the kernel density estimation of a given dataset and
returns a data frame with density values.
}
