% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_polar2.R
\name{coord_polar2}
\alias{coord_polar2}
\title{coord_polar2}
\usage{
coord_polar2(
  theta = "x",
  r0 = 0,
  r1 = 1,
  start = 0,
  end = 360,
  direction = 1,
  clip = "off",
  nice.facing = FALSE,
  x.label.rot = FALSE,
  x.label.pos = c("top", "bottom"),
  x.label.shift = 0.01,
  y.label.show = FALSE,
  y.label.pos = c("left", "right"),
  y.label.shift = 1
)
}
\arguments{
\item{theta}{The variable to be mapped to the angle (x or y). Default is "x".}

\item{r0}{The inner radius of the polar coordinate system. Default is 0.}

\item{r1}{The outer radius of the polar coordinate system. Default is 1.}

\item{start}{The starting angle in degrees. Default is 0.}

\item{end}{The ending angle in degrees. Default is 360.}

\item{direction}{The direction of rotation (1 for counterclockwise, -1 for clockwise). Default is 1.}

\item{clip}{Clip points outside the polar coordinate system (on/off). Default is "off".}

\item{nice.facing}{Should facing labels be automatically adjusted for clarity? Default is FALSE.}

\item{x.label.rot}{Should x-axis labels be rotated? Default is FALSE.}

\item{x.label.pos}{The position of x-axis labels (top or bottom). Default is "top".}

\item{x.label.shift}{The shift of x-axis labels. Default is 0.01.}

\item{y.label.show}{Show y-axis labels (TRUE/FALSE). Default is FALSE.}

\item{y.label.pos}{The position of y-axis labels (left or right). Default is "left".}

\item{y.label.shift}{The shift of y-axis labels. Default is 1.}
}
\value{
A ggplot2 polar coordinate system.
}
\description{
Create a polar coordinate system.
}
\details{
This function creates a polar coordinate system for a ggplot2 plot.
}
