% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_annosegment.R
\name{create_annosegment}
\alias{create_annosegment}
\title{Create an annotation segment for dotplot.}
\usage{
create_annosegment(
  data = NULL,
  panel_params = NULL,
  branch.side = c("top", "right", "bottom", "left"),
  branch.shift = 0,
  branch.len = 0.9,
  branch.height = 0.05,
  branch.lwd = 0.5,
  branch.label.size = 10,
  branch.label.shift = 0.2
)
}
\arguments{
\item{data}{A data frame containing the data for annotation segments.}

\item{panel_params}{A set of panel parameters (optional).}

\item{branch.side}{The side of the annotation segment (one of "top", "right", "bottom", or "left").}

\item{branch.shift}{The shift of the annotation segment along its side.}

\item{branch.len}{The length of the annotation segment.}

\item{branch.height}{The height of the annotation segment.}

\item{branch.lwd}{The line width of the annotation segment.}

\item{branch.label.size}{The font size of labels on the annotation segment.}

\item{branch.label.shift}{The shift of labels from the annotation segment.}
}
\value{
A list of graphical objects representing the annotation segment.
}
\description{
This function generates annotation segments for use in ggplot2 plots.
It allows you to add custom annotation segments to your plot.
}
