% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dendrogrob.R
\name{create_dendrogrob}
\alias{create_dendrogrob}
\title{Create a Dendrogram-based Grob}
\usage{
create_dendrogrob(
  data = NULL,
  panel_params = NULL,
  exp_mat = NULL,
  tree.type = c("rectangle", "triangle"),
  add.tree.y = FALSE,
  tree.y.side = "left",
  tree.y.width = 0.06,
  add.tree.x = FALSE,
  tree.x.side = "bottom",
  tree.x.height = 0.06,
  tree.y.label.hjust = 1,
  tree.x.label.hjust = 1,
  tree.x.shift = 0,
  tree.y.shift = 0,
  new.ylabel.x = -0.05,
  new.ylabel.width = 0.025,
  new.ylabel.size = 8,
  new.ylabel.face = "plain",
  new.ylabel.rot = 0,
  new.xlabel.y = -0.05,
  new.xlabel.height = 0.025,
  new.xlabel.size = 8,
  new.xlabel.face = "plain",
  new.xlabel.rot = 90
)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{panel_params}{A list of panel parameters.}

\item{exp_mat}{Matrix data with values.}

\item{tree.type}{Type of dendrogram tree ("rectangle" or "triangle").}

\item{add.tree.y}{Whether to add a vertical dendrogram.}

\item{tree.y.side}{Side for the vertical dendrogram ("left" or "right").}

\item{tree.y.width}{Width of the vertical dendrogram.}

\item{add.tree.x}{Whether to add a horizontal dendrogram.}

\item{tree.x.side}{Side for the horizontal dendrogram ("bottom" or "top").}

\item{tree.x.height}{Height of the horizontal dendrogram.}

\item{tree.y.label.hjust}{Horizontal justification for y-axis labels (0 or 1).}

\item{tree.x.label.hjust}{Horizontal justification for x-axis labels (0 or 1).}

\item{tree.x.shift}{Shift for the dendrogram placement.}

\item{tree.y.shift}{Shift for the dendrogram placement.}

\item{new.ylabel.x}{X-coordinate for the new y-axis labels.}

\item{new.ylabel.width}{Width of the new y-axis labels.}

\item{new.ylabel.size}{Font size for the new y-axis labels.}

\item{new.ylabel.face}{Font face for the new y-axis labels.}

\item{new.ylabel.rot}{Rotation angle for the new y-axis labels.}

\item{new.xlabel.y}{Y-coordinate for the new x-axis labels.}

\item{new.xlabel.height}{Height of the new x-axis labels.}

\item{new.xlabel.size}{Font size for the new x-axis labels.}

\item{new.xlabel.face}{Font face for the new x-axis labels.}

\item{new.xlabel.rot}{Rotation angle for the new x-axis labels.}
}
\value{
A list of graphical objects including the vertical and horizontal dendrograms
and their associated axis labels.
}
\description{
This function creates a dendrogram-based graphical object (Grob) with optional
vertical and horizontal dendrograms on the plot, along with associated axis labels.
}
