% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_line2.R
\name{element_line2}
\alias{element_line2}
\title{element_line2}
\usage{
element_line2(
  colour = NULL,
  linewidth = NULL,
  linetype = NULL,
  lineend = NULL,
  color = NULL,
  arrow = NULL,
  inherit.blank = FALSE,
  size = lifecycle::deprecated(),
  x.line.st = 0,
  x.line.ed = 1,
  y.line.st = 0,
  y.line.ed = 1,
  label.shift = -0.1,
  label = "dim",
  fontface = "bold.italic",
  fontsize = 8
)
}
\arguments{
\item{colour}{colour}

\item{linewidth}{linewidth}

\item{linetype}{Line type. An integer (0:8), a name (blank, solid,
dashed, dotted, dotdash, longdash, twodash), or a string with
an even number (up to eight) of hexadecimal digits which give the
lengths in consecutive positions in the string.}

\item{lineend}{lineend}

\item{color}{color}

\item{arrow}{Arrow specification, as created by \code{\link[grid:arrow]{grid::arrow()}}}

\item{inherit.blank}{inherit.blank}

\item{size}{size}

\item{x.line.st}{x line start position default(0).}

\item{x.line.ed}{x line end position default(1).}

\item{y.line.st}{y line start position default(0).}

\item{y.line.ed}{y line end position default(1).}

\item{label.shift}{label distance to the line default(-0.1).}

\item{label}{labels around the line default("dim").}

\item{fontface}{label fontface default("bold.italic").}

\item{fontsize}{label fontsize default(8).}
}
\description{
element_line2 control the axis line length
}
