% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-packs.R
\name{facet_feature}
\alias{facet_feature}
\title{Create custom facet wrap for features}
\usage{
facet_feature(
  facets = "gene_name",
  ncol = 4,
  scales = "free",
  strip.col = NULL,
  ...
)
}
\arguments{
\item{facets}{A character vector specifying the facet variable(s).}

\item{ncol}{The number of columns in the facet grid (default is 4).}

\item{scales}{Scales argument for facet_wrap (default is "free").}

\item{strip.col}{Background color for facet strip labels (optional).}

\item{...}{Additional arguments passed to ggh4x::facet_wrap2.}
}
\value{
A custom facet wrap for feature-based plotting.
}
\description{
This function creates a custom facet wrap for features in ggplot2.
}
