% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-packs.R
\name{facet_hetamap}
\alias{facet_hetamap}
\title{Create custom facet grid for hetamap plotting}
\usage{
facet_hetamap(
  facet_col = NULL,
  facet_row = NULL,
  scales = "free_x",
  strip.col = NULL,
  space = "free_x",
  no.xlabel = TRUE,
  ...
)
}
\arguments{
\item{facet_col}{A character vector specifying the columns for facetting (optional).}

\item{facet_row}{A character vector specifying the rows for facetting (optional).}

\item{scales}{Scales argument for facet_grid2 (default is "free_x").}

\item{strip.col}{Background color for facet strip labels (optional).}

\item{space}{Space argument for facet_grid2 (default is "free_x").}

\item{no.xlabel}{Whether remove x axis labels (default is "TRUE").}

\item{...}{Additional arguments passed to ggh4x::facet_grid2.}
}
\value{
A custom facet grid for hetamap plotting.
}
\description{
This function creates a custom facet grid for hetamap plotting in ggplot2.
}
