% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-packs.R
\name{facet_new}
\alias{facet_new}
\title{Custom ggplot2 facet function: facet_new}
\usage{
facet_new(
  facet_col = NULL,
  facet_row = NULL,
  scales = "free_x",
  strip.col = NULL,
  space = "free_x",
  no.xlabel = FALSE,
  no.ylabel = FALSE,
  title.hjust = 0,
  t = 0.025,
  r = 0.1,
  b = 0.025,
  l = 0.1,
  low.col = "grey90",
  high.col = "red",
  x.angle = 0,
  x.label.hjust = 0.5,
  legend.position = NULL,
  ...
)
}
\arguments{
\item{facet_col}{Columns to facet by (can be multiple columns).}

\item{facet_row}{Rows to facet by (can be multiple rows).}

\item{scales}{The type of scaling to use for facets (default is "free_x").}

\item{strip.col}{Background color for facet strip labels.}

\item{space}{The type of spacing to use for facets (default is "free_x").}

\item{no.xlabel}{Logical value indicating whether to hide x-axis labels (default is FALSE).}

\item{no.ylabel}{Logical value indicating whether to hide y-axis labels (default is FALSE).}

\item{title.hjust}{Horizontal justification for facet titles (0 for left, 0.5 for center, 1 for right).}

\item{t}{Top margin size for the plot (in npc units).}

\item{r}{Right margin size for the plot (in npc units).}

\item{b}{Bottom margin size for the plot (in npc units).}

\item{l}{Left margin size for the plot (in npc units).}

\item{low.col}{Color for the low end of the gradient.}

\item{high.col}{Color for the high end of the gradient.}

\item{x.angle}{Angle for x-axis labels (default is 0).}

\item{x.label.hjust}{X axis label hjust (default is 0.5).}

\item{legend.position}{Position of the legend (e.g., "top", "bottom", "right", "none").}

\item{...}{Additional arguments to be passed to \code{\link[ggh4x]{facet_grid2}}.}
}
\value{
A list of ggplot2 elements, including facet_grid settings, theme settings,
}
\description{
This custom ggplot2 facet function, \code{facet_new}, allows you to create
customized facet grids in ggplot2, with additional options for appearance and scales.
}
