% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_wrapnew.R
\name{facet_wrapnew}
\alias{facet_wrapnew}
\title{Wrap a 1d ribbon of panels into 2d (slightly modified with facet_wrap)}
\usage{
facet_wrapnew(
  facets,
  nrow = NULL,
  ncol = NULL,
  scales = "fixed",
  shrink = TRUE,
  labeller = "label_value",
  as.table = TRUE,
  switch = deprecated(),
  drop = TRUE,
  dir = "h",
  strip.position = "top",
  axes = "margins",
  axis.labels = "all"
)
}
\arguments{
\item{facets}{A set of variables or expressions quoted by \code{\link[=vars]{vars()}}
and defining faceting groups on the rows or columns dimension.
The variables can be named (the names are passed to \code{labeller}).

For compatibility with the classic interface, can also be a
formula or character vector. Use either a one sided formula, \code{~a + b},
or a character vector, \code{c("a", "b")}.}

\item{nrow, ncol}{Number of rows and columns.}

\item{scales}{Should scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"})?}

\item{dir}{Direction: either \code{"h"} for horizontal, the default, or \code{"v"},
for vertical. When \code{"h"} or \code{"v"} will be combined with \code{as.table} to
set final layout. Alternatively, a combination of \code{"t"} (top) or
\code{"b"} (bottom) with \code{"l"} (left) or \code{"r"} (right) to set a layout directly.
These two letters give the starting position and the first letter gives
the growing direction. For example \code{"rt"} will place the first panel in
the top-right and starts filling in panels right-to-left.}

\item{strip.position}{By default, the labels are displayed on the top of
the plot. Using \code{strip.position} it is possible to place the labels on
either of the four sides by setting \code{strip.position = c("top",
  "bottom", "left", "right")}}

\item{axes}{Determines which axes will be drawn in case of fixed scales.
When \code{"margins"} (default), axes will be drawn at the exterior margins.
\code{"all_x"} and \code{"all_y"} will draw the respective axes at the interior
panels too, whereas \code{"all"} will draw all axes at all panels.}

\item{axis.labels}{Determines whether to draw labels for interior axes when
the scale is fixed and the \code{axis} argument is not \code{"margins"}. When
\code{"all"} (default), all interior axes get labels. When \code{"margins"}, only
the exterior axes get labels, and the interior axes get none. When
\code{"all_x"} or \code{"all_y"}, only draws the labels at the interior axes in the
x- or y-direction respectively.}
}
\description{
\code{facet_wrapnew()} wraps a 1d sequence of panels into 2d. This is generally
a better use of screen space than \code{\link[=facet_grid]{facet_grid()}} because most
displays are roughly rectangular.
}
