% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_markArrow.R
\name{geom_markArrow}
\alias{geom_markArrow}
\title{geom_markArrow}
\usage{
geom_markArrow(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  lineend = "butt",
  linejoin = "round",
  rel.pos = 0.1,
  rel.len = 0.3,
  arrow = grid::arrow(type = "closed", length = grid::unit(0.4, "cm")),
  arrow.fill = NULL,
  colour = "black",
  label = c("UMAP1", "UMAP2"),
  label.size = 3,
  label.shift = c(0.025, 0.025),
  fontface = "bold.italic",
  tail.shift = 0,
  corner.pos = "left_b",
  facet.vars = NULL,
  ...
)
}
\arguments{
\item{mapping}{The aesthetic mappings for the layer.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data.}

\item{position}{The method used to position the objects.}

\item{na.rm}{If \code{TRUE}, removes missing values from the data.}

\item{show.legend}{Logical. Should this layer be included in the legends? NA,
the default, includes if any aesthetics are mapped. FALSE never includes, and
TRUE always includes.}

\item{inherit.aes}{If \code{TRUE} (the default), the plot's aesthetics will
be inherited. Set to \code{FALSE} to create a completely new set of aesthetic
mappings.}

\item{lineend}{The line ending style of the path; one of "butt" (default),
"round", or "square".}

\item{linejoin}{The line join style of the path; one of "round" (default),
"mitre", or "bevel".}

\item{rel.pos}{A numeric value between 0 and 1 indicating where along the
length of the path the arrowhead should be placed.}

\item{rel.len}{A numeric value between 0 and 1 indicating the size of the
arrowhead relative to the length of the path.}

\item{arrow}{The arrowhead to use, created using grid::arrow() or grid::arrowGrob().}

\item{arrow.fill}{The fill color for the arrowhead.}

\item{colour}{The color of the line and arrowhead.}

\item{label}{A character vector of length 2 indicating the labels for the
start and end points of the line.}

\item{label.size}{The font size for the labels.}

\item{label.shift}{A numeric vector of length 2 indicating the x and y offsets
for the label positions.}

\item{fontface}{The font face for the labels.}

\item{tail.shift}{A numeric value indicating the offset for the start point
of the line along its direction.}

\item{corner.pos}{A character vector indicating the position of the label
corner relative to the path; one of "left_b", "right_b", "left_t", or "right_t".}

\item{facet.vars}{A list of variables to use for faceting.}

\item{...}{description}
}
\value{
An object of class \code{GeomMarkArrow}.
}
\description{
geom_markArrow
}
\author{
Jun Zhang

An extension of ggplot2's GeomPath that adds an arrowhead to the end of a line.
}
