% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_marklabel.R
\name{geom_marklabel}
\alias{geom_marklabel}
\title{geom_marklabel}
\usage{
geom_marklabel(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  mark.label = NULL,
  shift = 0,
  use.smartAlign2 = FALSE,
  link.line.length = 0.075,
  link.label.space = 0.025,
  mark.scale = c(0, 1),
  link.label.gp = gpar(fontsize = 10),
  link.line.gp = gpar(fill = "black", col = "black"),
  link.circle.start.gp = gpar(fill = "black", col = "black"),
  link.circle.end.gp = gpar(fill = "black", col = "black"),
  link.start.type = c("line", "circle", "arrow"),
  link.end.type = c("line", "circle", "arrow"),
  circle.arrow.size = c(0.01, 0.01),
  pos = c("right", "left", "top", "bottom")
)
}
\arguments{
\item{mapping}{Aesthetic mapping.}

\item{data}{The data to be plotted.}

\item{stat}{The statistical transformation to be applied to the data.}

\item{position}{The position adjustment method for overlapping points.}

\item{...}{Additional arguments to be passed to GeomMarklabel.}

\item{na.rm}{Logical. Should missing values be removed?}

\item{show.legend}{Should this layer be included in the legend?}

\item{inherit.aes}{Should inherit aesthetics from the parent plot?}

\item{mark.label}{The labels to be displayed with markers.}

\item{shift}{The shift of labels (0).}

\item{use.smartAlign2}{Logical, whether to use smart alignment.}

\item{link.line.length}{Length of the connecting lines.}

\item{link.label.space}{Spacing between the label and connecting lines.}

\item{mark.scale}{A numeric vector specifying the scale for mark positioning.}

\item{link.label.gp}{A graphical parameter for the label appearance.}

\item{link.line.gp}{A graphical parameter for the connecting lines.}

\item{link.circle.start.gp}{A graphical parameter for the start circle (if used).}

\item{link.circle.end.gp}{A graphical parameter for the end circle (if used).}

\item{link.start.type}{Type of shape at the start of the connecting lines ("line", "circle", or "arrow").}

\item{link.end.type}{Type of shape at the end of the connecting lines ("line", "circle", or "arrow").}

\item{circle.arrow.size}{A numeric vector specifying the size of circles or arrows.}

\item{pos}{Alignment position ("right", "left", "top", or "bottom").}
}
\value{
A ggplot2 layer for labeled markers.
}
\description{
Add labeled markers to a plot.
}
\details{
This function adds labeled markers to a plot using the GeomMarklabel layer.
}
