% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scPoint.R
\name{geom_scPoint}
\alias{geom_scPoint}
\title{Create a custom ggplot2 point geom}
\usage{
geom_scPoint(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Aesthetic mapping for the plot.}

\item{data}{A data frame containing the single-cell data.}

\item{stat}{The statistical transformation to apply (default is "identity").}

\item{position}{The positioning transformation to apply (default is "identity").}

\item{...}{Additional arguments passed to the geom.}

\item{na.rm}{Should missing values be removed (default is FALSE).}

\item{show.legend}{Should the legend be displayed (default is NA).}

\item{inherit.aes}{Should aesthetics be inherited (default is TRUE).}
}
\value{
A custom ggplot2.
}
\description{
This function creates a custom ggplot2 point geom.
}
