% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scPoint2.R
\name{geom_scPoint2}
\alias{geom_scPoint2}
\title{Create a custom ggplot2 geom for single-cell data point plotting with labels and legend}
\usage{
geom_scPoint2(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  label.gp = gpar(),
  add_label = TRUE,
  add_legend = FALSE,
  ncol = 1,
  vgap = 0.5,
  point.size = 2,
  point.label.size = 10,
  point.label.col = "black",
  lgd_x = 1.05,
  lgd_y = 0.5,
  lgd_width = 0.1,
  lgd_height = 0.9,
  fontsize = 10,
  fontface = "plain"
)
}
\arguments{
\item{mapping}{Aesthetic mapping for the plot.}

\item{data}{A data frame containing the single-cell data.}

\item{stat}{The statistical transformation to apply (default is "identity").}

\item{position}{The positioning transformation to apply (default is "identity").}

\item{...}{Additional arguments passed to the geom.}

\item{na.rm}{Should missing values be removed (default is FALSE).}

\item{show.legend}{Should the legend be displayed (default is NA).}

\item{inherit.aes}{Should aesthetics be inherited (default is TRUE).}

\item{label.gp}{A list of graphical parameters for labels (optional).}

\item{add_label}{Should labels be added to the points (default is TRUE).}

\item{add_legend}{Should a new legend be added (default is FALSE).}

\item{ncol}{Number of columns in the legend (default is 1).}

\item{vgap}{Vertical gap between legend items (default is 0.5).}

\item{point.size}{Size of legend points (default is 2).}

\item{point.label.size}{Size of legend point labels (default is 10).}

\item{point.label.col}{Color of legend point labels (default is "black").}

\item{lgd_x}{X-position of the legend (default is 1.05).}

\item{lgd_y}{Y-position of the legend (default is 0.5).}

\item{lgd_width}{Width of the legend (default is 0.1).}

\item{lgd_height}{Height of the legend (default is 0.9).}

\item{fontsize}{Font size for labels and legend (default is 10).}

\item{fontface}{Font face for labels and legend (default is "plain").}

\item{hgap}{Horizontal gap between legend items (default is 1).}
}
\value{
A custom ggplot2 geom for single-cell data point plotting with labels and legend.
}
\description{
This function creates a custom ggplot2 geom for single-cell data point plotting with labels and optional legend.
}
