% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscplot.R
\name{ggscplot}
\alias{ggscplot}
\title{Create a ggplot for single-cell data visualization}
\usage{
ggscplot(
  data = NULL,
  mapping = aes(),
  object = NULL,
  reduction = "umap",
  features = NULL,
  featuresAnno = 0,
  pct.exp.var = "seurat_clusters",
  slot = "data",
  environment = parent.frame(),
  ...
)
}
\arguments{
\item{data}{A data frame containing the single-cell data.}

\item{mapping}{Aesthetic mapping for the plot.}

\item{object}{A Seurat object (optional) from which to fetch data.}

\item{reduction}{The reduction method to use (default is "umap").}

\item{features}{A vector of features to include in the plot (optional).}

\item{featuresAnno}{Annotation level for features (default is 0).}

\item{pct.exp.var}{A variable for percentage of expression (default is "seurat_clusters").}

\item{slot}{The data slot to use from the Seurat object (default is "data").}

\item{environment}{environment.}

\item{...}{Additional arguments passed to ggplot.}
}
\value{
A ggplot object.
}
\description{
This function creates a ggplot object for visualizing single-cell data which extract relatted
data from Seurat object before plotting.
}
\seealso{
\code{\link{fetch_data}}
}
\concept{ggscplot functions}
