% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendGrob2.R
\name{legendGrob2}
\alias{legendGrob2}
\title{Create a custom legend grob (slightly modified from grid::legendGrob)}
\usage{
legendGrob2(
  labels = NULL,
  labels.point = NULL,
  labels.point.gp = gpar(),
  nrow = NULL,
  ncol = NULL,
  byrow = FALSE,
  do.lines = has.lty || has.lwd,
  lines.first = TRUE,
  hgap = unit(1, "lines"),
  vgap = unit(1, "lines"),
  default.units = "lines",
  point.size = 2,
  pch = NULL,
  gp = gpar(),
  vp = NULL
)
}
\arguments{
\item{labels}{A list of labels for the legend items.}

\item{labels.point}{A list of labels for legend point items (optional).}

\item{labels.point.gp}{A list of graphical parameters for point labels (optional).}

\item{nrow}{Number of rows in the legend (optional).}

\item{ncol}{Number of columns in the legend (optional).}

\item{byrow}{Should legend items be arranged by row (default is FALSE).}

\item{do.lines}{Should lines be included in the legend (default is TRUE if 'lty' or 'lwd' are specified in 'gp', otherwise FALSE).}

\item{lines.first}{Should lines be placed before points in the legend (default is TRUE).}

\item{hgap}{Horizontal gap between legend items (default is 1 line).}

\item{vgap}{Vertical gap between legend items (default is 1 line).}

\item{default.units}{Default units for gaps (default is "lines").}

\item{point.size}{Size of legend points (default is 2).}

\item{pch}{A vector of point symbols (optional).}

\item{gp}{A list of graphical parameters for lines and borders (optional).}

\item{vp}{A grid viewport in which to draw the legend (optional).}
}
\value{
A custom legend grob.
}
\description{
This function creates a custom legend grob, allowing for
fine-grained control over the appearance of legend items.
}
