% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sc.R
\name{theme_sc}
\alias{theme_sc}
\title{Customize ggplot2 theme for single-cell plots}
\usage{
theme_sc(
  x.line.len = 0.25,
  x.label = "dim 1",
  y.line.len = 0.25,
  y.label = "dim 1",
  label.shift = -2.5,
  fontsize = 10,
  border.col = NA,
  add.arrow = TRUE,
  ...
)
}
\arguments{
\item{x.line.len}{Length of the x-axis arrow (default is 0.25).}

\item{x.label}{Label for the x-axis (default is "dim 1").}

\item{y.line.len}{Length of the y-axis arrow (default is 0.25).}

\item{y.label}{Label for the y-axis (default is "dim 1").}

\item{label.shift}{Shift for axis labels (default is -2.5).}

\item{fontsize}{Font size for labels (default is 10).}

\item{border.col}{Color of the plot border (default is NA).}

\item{add.arrow}{Should arrows be added to the axis lines? (default is TRUE).}

\item{...}{Additional arguments passed to ggplot2 theme.}
}
\value{
A ggplot2 theme object customized for single-cell plots.
}
\description{
This function customizes the ggplot2 theme for single-cell plots by adjusting axis lines, labels, fonts, background color, and more.
}
