% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscplot.R
\name{ggscplot}
\alias{ggscplot}
\title{Create a ggplot for single-cell data visualization}
\usage{
ggscplot(
  data = NULL,
  mapping = aes(),
  object = NULL,
  reduction = "umap",
  features = NULL,
  featuresAnno = 0,
  pct.exp.var = "seurat_clusters",
  slot = "data",
  gene.order = NULL,
  group.vars = NULL,
  t = 0.1,
  r = 0.1,
  b = 0.1,
  l = 0.1,
  environment = parent.frame(),
  ...
)
}
\arguments{
\item{data}{A data frame containing the single-cell data.}

\item{mapping}{Aesthetic mapping for the plot.}

\item{object}{A Seurat object (optional) from which to fetch data.}

\item{reduction}{The reduction method to use (default is "umap").}

\item{features}{A vector of features to include in the plot (optional).}

\item{featuresAnno}{Annotation level for features (default is 0).}

\item{pct.exp.var}{A variable for percentage of expression (default is "seurat_clusters").}

\item{slot}{The data slot to use from the Seurat object (default is "data").}

\item{gene.order}{Character vectors to reorder the gene labels.}

\item{group.vars}{The column names in metadata which you want to combine,
useful for facet plot, the new combined column name is "group.vars" and value
column name is "group.value" (default is NULL).}

\item{t}{the top margin of the plot (default is 0.1 npc).}

\item{r}{the right margin of the plot (default is 0.1 npc).}

\item{b}{the bottom margin of the plot (default is 0.1 npc).}

\item{l}{the left margin of the plot (default is 0.1 npc).}

\item{environment}{The levels of the gene name.}

\item{...}{Additional arguments passed to ggplot.}
}
\value{
A ggplot object.
}
\description{
This function creates a ggplot object for visualizing single-cell data which extract relatted
data from Seurat object before plotting.
}
\seealso{
\code{\link{fetch_data}}
}
\concept{ggscplot functions}
